cast division to int

+++ src/process_pages.py
@@ -44,7 +44,7 @@ def process_seed_pages(pages_db, redirects_lookup, seed_pages, depth=1):
     print("Fetching and processing", len(page_titles_to_fetch), "pages in", batch_size, "batches")
     for batch_num, titles_batch in progressbar(enumerate(u.create_batches(list(page_titles_to_fetch),
                                                                           batch_size=batch_size)),
-                                               max_value=len(page_titles_to_fetch)/batch_size):
+                                               max_value=int(len(page_titles_to_fetch)/batch_size)):
       batch_pages_to_process = _fetch_pages(pages_db, titles_batch)
       latest_processed_pages = _process_pages(redirects_lookup, batch_pages_to_process)
       processed_pages += latest_processed_pages
